Improve run_multiple_simulations
Note: This is a breaking change so we also modify the relevant scripts
that call this function, In particular, we no longer rely on a
`num_simulations` parameter to determine how many simulations to run.
Now, we usually pass a dict containing a vector of seeds or pass a list
of dicts.

Add dict_list to utils.py

Create new default for simulation_id in DebateManager

Add networks.py

Initialize graph in DebateManager

Note: This is a breaking change so we also remove the code for
initializing graphs from our scripts.

Collect graphs in run_multiple_simulations

Add serialize_graphs method to data_utils.py

Serialize dictionaries of networkx graphs

Compute graph_metrics in scripts